Update Wallet entity and dependent classes to use long values instead of double; Add locks to transactions

+++ bank/src/main/java/ru/duckcoder/bankservice/dto/wallet/WalletDTO.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class WalletDTO {
-    private Double deposit;
-    private Double accrual;
+    private Long deposit;
+    private Long accrual;
 }
